requirements.txt
asgiref==3.9.1
attrs==25.3.0
autobahn==24.4.2
Automat==25.4.16
cffi==1.17.1
channels==4.3.1
constantly==23.10.4
cryptography==45.0.5
daphne==4.2.1
Django==5.2.4
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
numpy==2.3.2
pandas==2.3.1
pillow==11.3.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pyOpenSSL==25.1.0
python-dateutil==2.9.0.post0
pytz==2025.2
service-identity==24.2.0
setuptools==80.9.0
six==1.17.0
sqlparse==0.5.3
Twisted==25.5.0
txaio==25.6.1
typing_extensions==4.14.1
tzdata==2025.2
zope.interface==7.2
python-dotenv==1.0.0
dj-database-url==2.1.0
channels-redis==4.2.0
redis==5.0.1
psycopg2-binary==2.9.9
gunicorn==21.2.0
whitenoise==6.6.0
